+++ packages/core/lib/index.ts
@@ -1,20 +1,54 @@
-import './index.css';
-import { Display } from './components/Display';
-import { Heading } from './components/Heading';
-import { Title } from './components/Title';
-import { Body } from './components/Body';
-import { Detail } from './components/Detail';
-import { Label } from './components/Label';
-import { Link } from './components/Link';
-import { colors } from './colors/color';
+import "./index.css";
+import { Display } from "./components/Display";
+import { Heading } from "./components/Heading";
+import { Title } from "./components/Title";
+import { Body } from "./components/Body";
+import { Detail } from "./components/Detail";
+import { Label } from "./components/Label";
+import { Link } from "./components/Link";
+import { colors } from "./colors/color";
 
-import { Button } from './components/Button';
-import { LinkButton } from './components/LinkButton';
-import { Tag } from './components/Tag';
-import { Spinner } from './components/Spinner';
-import { Badge } from './components/Badge';
-import { Select } from './components/Select';
+import { Button } from "./components/Button";
+import { LinkButton } from "./components/LinkButton";
+import { Tag } from "./components/Tag";
+import { Spinner } from "./components/Spinner";
+import { Badge } from "./components/Badge";
+import { TextInput } from "./components/TextInput";
+import { TextArea } from "./components/TextArea";
+import { Breadcrumb } from "./components/Breadcrumb";
+import { Switch } from "./components/Switch";
+import { Chip } from "./components/Chip";
+import { Checkbox } from "./components/Checkbox";
+import { RadioButtonGroup } from "./components/RadioButton";
+import { Tabs } from "./components/Tab";
+import { Modal } from "./components/Modal";
+import { Accordion } from "./components/Accordion";
+import { Disclosure } from "./components/Disclosure";
+import { Pagination } from "./components/Pagination";
+import { FileUpload } from "./components/FileUpload";
+import { Calendar } from "./components/Calendar";
+import { Select } from "./components/Select";
 
 export { Display, Heading, Title, Body, Detail, Label, Link, colors };
-export { Button, LinkButton, Tag, Spinner, Badge };
-export { Select };
+export {
+  Button,
+  LinkButton,
+  Tag,
+  Spinner,
+  Badge,
+  Switch,
+  Chip,
+  Checkbox,
+  RadioButtonGroup,
+  TextInput,
+  TextArea,
+  Breadcrumb,
+  Tabs,
+  Modal,
+  Accordion,
+  Disclosure,
+  Pagination,
+  FileUpload,
+  Calendar,
+  Select,
+};